Main item 25 spindle turns off prematurely in feedhold (#28)
Pushing fix for part 2 of issue 2. This should cause the spindle to stay on until Z-lift is either done or nearly done.

This should fix the issue where the spindle turns off before the Z-lift has started.
It should now cause the spindle to stay on until Z-lift is either done or nearly done.

See code review approval for tests performed.