Reverting some changes due to new fix
Reverting back to just the basic fix that fixes the main issue in #24 as in exploring the old issue #11 we found a better more elegant and comprehensive way to fix the SR issue. So a small SR error will be introduced in this fix, but will be fixed in a sister pull request.