Main item 24 feedhold resume displays out1 aka spindle in sr after dwell (#30)
* Initial fix for issue #24

This fixes the primary issue with the spindle "out:1" status not displaying until after the dwell. However it introduces a new (but not totally new) bug where it shoes the current Z position as being at its final destination instead of its current destination. Working to fix that next...

* First attempt at SR filtering on pos, eats SRs

* Incremental progress..

Refined the fix a bit, making it more accurate. Still has issues with eating some extra SR's and making the plunge/withdrawl of the spindle seem instant in SR's.

* Cleanup position check, make safer

* Add check for MOTION_STOP (MODEL runtime) and cleanup, seems to run better

* Pushing cleaned up fix for #24

This should fully fix issue #24 and produce accurate SR's for it.

* Reverting some changes due to new fix

Reverting back to just the basic fix that fixes the main issue in #24 as in exploring the old issue #11 we found a better more elegant and comprehensive way to fix the SR issue. So a small SR error will be introduced in this fix, but will be fixed in a sister pull request.

Co-authored-by: Matt Staniszewski <matt@msxconsulting.com>