Main item 34 faulty SRs from feedhold or kill (#35)
* Initial commit for old item #11 fix

This fixes the issue where one of the SR's that comes out when you feedhold or kill shows the target instead of the current position.

* Fix 2

This also fixes the issue with the SR's, but may be more comprehensive and thorough in it's approach. May reduce the SRs produced slightly in a g38.3

* Fix 2.1

Part of "Fix 2" commit, missed a line.

* Cleaning up fix for #34

This fixes the faulty SR's that show target instead of current value when feedholding or issuing a kill. Happens in at least G0, G1 and G38.3